Add apkinfo man page
Debian requires all commands to have man pages; as such, I've been
including a man page in the 'apkinfo' Debian package for quite some
time. With this commit, it's cleaned up a little bit and ready for
upstream inclusion.

License for the man page is the same as the rest of the package
(Apache-2.0).